+++ lib/src/components/buttons/primary.dart
@@ -4,7 +4,7 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 
 import '../../../flutter_dsfr.dart';
-import 'base_button.dart';
+import 'base.dart';
 
 /// Specs: https://gouvfr.atlassian.net/wiki/spaces/DB/pages/217284660/Boutons+-+Buttons#Bouton-primaire
 class DSFRPrimaryButton extends DSFRGroupeableButton {
