+++ lib/src/components/buttons/secondary.dart
@@ -3,36 +3,34 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 
+import '../../theme/button_style.dart';
 import '../../theme/colors.dart';
 import '../../theme/sizes.dart';
 import '../../theme/typography.dart';
-import 'base_button.dart';
+import 'base.dart';
 
-/// Specs: https://gouvfr.atlassian.net/wiki/spaces/DB/pages/217284660/Boutons+-+Buttons#Bouton-tertiaire
-class DSFRTertiaryButton extends DSFRBaseButton {
-  final bool noOutline;
-
-  const DSFRTertiaryButton({
+class DSFRSecondaryButton extends DSFRGroupeableButton {
+  const DSFRSecondaryButton({
     super.key,
     required super.onPressed,
     required super.label,
     super.icon,
-    super.iconPosition,
-    this.noOutline = false,
+    super.iconPosition = IconPosition.left,
+    super.mainAxisSize,
   });
 
-  const DSFRTertiaryButton.icon({
+  const DSFRSecondaryButton.icon({
     super.key,
-    required super.icon,
     required super.onPressed,
-    this.noOutline = false,
+    required super.icon,
   }) : super.icon();
 
   @override
   Widget build(BuildContext context) {
-    final dsfrColors = DSFRColors.of(context);
     final dsfrSpacings = DSFRSizes.of(context);
     final dsfrTypography = DSFRTypography.of(context);
+    final dsfrColors = DSFRColors.of(context);
+    final dsfrButtonStyle = DSFRButtonStyle.of(context);
 
     final scale = MediaQuery.maybeOf(context)?.textScaleFactor ?? 1;
     final double gap =
@@ -45,17 +43,14 @@ class DSFRTertiaryButton extends DSFRBaseButton {
         : dsfrColors.textDisabledGrey;
 
     return RawMaterialButton(
+      onPressed: onPressed,
       elevation: 0.0,
       disabledElevation: 0.0,
       fillColor: Colors.transparent,
+      hoverColor: dsfrColors.blockColorHover,
+      splashColor: dsfrColors.blockColorActive,
       shape: RoundedRectangleBorder(
-        side: !noOutline
-            ? BorderSide(
-                color: enabled
-                    ? dsfrColors.defaultBorderGrey
-                    : dsfrColors.textDisabledGrey,
-              )
-            : BorderSide.none,
+        side: BorderSide(color: foregroundColor, width: 1),
       ),
       padding: EdgeInsets.symmetric(
         vertical: dsfrSpacings.w1,
@@ -63,12 +58,11 @@ class DSFRTertiaryButton extends DSFRBaseButton {
       ),
       constraints: const BoxConstraints(),
       textStyle: dsfrTypography.btnLabel.copyWith(color: foregroundColor),
-      onPressed: onPressed,
       child: iconOnly
           ? btnIcon
           : Row(
               mainAxisAlignment: MainAxisAlignment.center,
-              mainAxisSize: MainAxisSize.min,
+              mainAxisSize: mainAxisSize ?? dsfrButtonStyle.mainAxisSize,
               children: [
                 if (btnIcon != null && iconPosition == IconPosition.left) ...[
                   btnIcon,
